chore(engine): move ZeebeState to immutable package
