+++ engine/src/main/java/io/zeebe/engine/processing/deployment/DeploymentCreateProcessor.java
@@ -26,9 +26,9 @@ import io.zeebe.engine.processing.streamprocessor.writers.TypedResponseWriter;
 import io.zeebe.engine.processing.streamprocessor.writers.TypedStreamWriter;
 import io.zeebe.engine.processing.streamprocessor.writers.Writers;
 import io.zeebe.engine.state.KeyGenerator;
-import io.zeebe.engine.state.ZeebeState;
 import io.zeebe.engine.state.immutable.ProcessState;
 import io.zeebe.engine.state.immutable.TimerInstanceState;
+import io.zeebe.engine.state.immutable.ZeebeState;
 import io.zeebe.engine.state.instance.TimerInstance;
 import io.zeebe.model.bpmn.util.time.Timer;
 import io.zeebe.protocol.impl.record.value.deployment.DeploymentRecord;
